// src/types/community.type.ts
export interface ICommunity {
  _id: string
  state: string
  region: string
  name: string
  censusCode: string
  cencusName: string
  remoteArea: string
  longitude: string
  latitude: string
  locationType: string
  area: string
  waterSource: string
  waterProvider: string
  dataSource: string
  progress?: string
  dataPeriod?: string
}

// export type